added arg  parsing

+++ test.sh
@@ -15,11 +15,11 @@
 # =============================== Global variables =========================== #
 # = = = = = = = = = = = = = = = YOU CAN TWEAK THESES = = = = = = = = = = = = = #
 # Specify the verbose level : 0 = pretty , 1 = minimal, 2 = full info 
-_VERBOSE=0
+_VERBOSE=2
 # The time out for execution of tests
-_TIME_OUT=5
+_TIME_OUT=1
 # The global max for fast tweaking of below maximums
-_GLOBAL_MAX_=10
+_GLOBAL_MAX_=20
 # 1 = add newlines add '\n' between each items in the sequence
 WITH_NEWLINES=0
 # Max number of generated chars
@@ -43,9 +43,6 @@ WD=.42_Test_42_Logs_Here        # The 'Working Directory'
 LOG_DIR=$WD/42.LOGS_HERE       	# The LOGS Directory
 LOG_FILE=$LOG_DIR/42.LOGS      	# The LOGS file
 LOG_DISPLAY_FILE=$WD/42.KO_LOGS.42.sh
-# Here you can see how the script searches for your input files
-FT_PRINTF_HEADER_FILE=$( mkdir -p $WD && cd $WD && find ../ -type f -name "ft_printf.h" )
-FT_PRINTF_LIB_FILE=$( find . -name "libftprintf.a" )
 # Down here are all the file names WITH 42's E V E R Y W H E R E
 no_dir_header_file=42.HeAdEr.h
 header_file=$WD/$no_dir_header_file
@@ -250,49 +247,89 @@ function write_sequence ()
     echo "#endif" >> $header_file
 }
 
-function timeout() { perl -e 'alarm shift; exec @ARGV' "$@"; }
+function time_out_kill()
+{
+    trap - ALRM TERM
+    kill -ALRM $sleepkill 2>/dev/null
+    kill -ALRM $! 2>/dev/null && return 124
+}
+
+function time_out_sleepkill()
+{
+    trap "time_out_kill" ALRM
+    sleep $1& wait
+    kill -TERM $command 2>/dev/null
+}
+
+function time_out ()
+{   
+    time_out_sleepkill $1& sleepkill=$!
+    shift
+    trap "time_out_kill" ALRM TERM
+    "$@" & command=$! ; wait $command; RET=$?
+    kill -TERM $sleepkill 2>/dev/null
+    return $RET          
+}
+
+function compile_run()
+{
+    rm -f $printf_exec_file
+    if [[ -e $printf_main_file ]] ; then
+        gcc -I. $printf_main_file -o $printf_exec_file
+    fi
+    if [[ -e $printf_exec_file ]] && time_out $_TIME_OUT ./$printf_exec_file > $printf_diff_file
+    then
+        echo >> $printf_diff_file
+    else
+        if [[ $_VERBOSE -ge 1 ]] ; then
+            printf "\nTEST : %-6d \033[31;1m STDIO PRINTF TIME OUT or EXEC ERROR\033[m\n" $test_n
+            echo $macro
+            KO_NUM=$((KO_NUM + 1))
+        fi
+        printf "\n= = = TEST : %-6d STDIO PRINTF TIMEOUT or EXEC ERROR\n" $test_n >> $LOG_FILE"_"$test_n
+		echo $macro >> $LOG_FILE"_"$test_n
+        return 2
+    fi
+    
+    rm -f $ft_printf_exec_file
+    if [[ -e $ft_printf_main_file ]] ; then
+       if ! gcc -I../ $ft_printf_main_file $FT_PRINTF_LIB_FILE -o $ft_printf_exec_file ; then 
+        printf "\033[1;31mCOMPILE ERROR\033[m\n" ; return 3
+        fi
+    fi
+    if [[ -e $ft_printf_exec_file ]] && time_out $_TIME_OUT ./$ft_printf_exec_file > $ft_printf_diff_file
+    then
+        echo >> $ft_printf_diff_file
+    else
+        if [[ $_VERBOSE -ge 1 ]]
+        then
+            printf "\nTEST : %-6d \033[31;1m FT_printf TIME OUT or EXEC ERROR\033[m\n" $test_n
+            echo $macro
+            KO_NUM=$((KO_NUM + 1))
+        fi
+        printf "\n= = = TEST : %-6d FT_printf TIME OUT or EXEC ERROR\n" $test_n >> $LOG_FILE"_"$test_n
+		echo $macro >> $LOG_FILE"_"$test_n
+        exit 2
+    fi
+}
 
 function run_test ()
 {
     write_header_file
-    rm -f $printf_exec_file
-    gcc -I../ $printf_main_file -o $printf_exec_file
-    if timeout $_TIME_OUT ./$printf_exec_file > $printf_diff_file
-    then
-        echo >> $printf_diff_file
-    else
-        printf "\rTEST : %-6d \033[31;1m STDIO PRINTF TIME OUT or EXEC ERROR\033[m\n" $test_n
-        printf "\n= = = TEST : %-6d STDIO PRINTF TIMEOUT or EXEC ERROR\n" \
-$test_n >> $LOG_FILE"_"$test_n
-		echo $macro >> $LOG_FILE"_"$test_n
-		cat $LOG_FILE"_"$test_n
-        exit 1
-    fi
-    rm -f $ft_printf_exec_file
-    if ! gcc -I../ $ft_printf_main_file $FT_PRINTF_LIB_FILE -o $ft_printf_exec_file
-    then 
-        printf "\033[1;31mCOMPILE ERROR\033[m\n"
-        exit 1
-    fi
-    if timeout $_TIME_OUT ./$ft_printf_exec_file > $ft_printf_diff_file
-    then
-        echo >> $ft_printf_diff_file
-    else
-        printf "\rTEST : %-6d \033[31;1m FT_printf TIME OUT or EXEC \
-ERROR\033[m\n" $test_n
-        printf "\n= = = TEST : %-6d FT_printf TIME OUT or EXEC \
-ERROR\n" $test_n >> $LOG_FILE"_"$test_n
-		echo $macro >> $LOG_FILE"_"$test_n
-        return 1
-    fi
+    
+    if ! compile_run ; then return 1
+    fi   
+
     printf "\n= = = TEST : %-6d = =\n" $test_n >> $LOG_FILE"_"$test_n
     echo $macro >> $LOG_FILE"_"$test_n
-	if [[ $_VERBOSE -eq 1 ]]
-	then 
-		printf "TEST% 6d : " $test_n
+	
+    if [[ $_VERBOSE -ge 1 ]]; then printf "TEST% 6d : " $test_n
 	fi
-    diff -a -s -u --label FT_42 $ft_printf_diff_file\
- --label STDIO $printf_diff_file >> $LOG_FILE"_"$test_n
+    
+    if [[ -e $ft_printf_diff_file ]] && [[ -e $printf_diff_file ]]; then
+        diff -a -s -u --label FT_42 $ft_printf_diff_file --label STDIO $printf_diff_file >> $LOG_FILE"_"$test_n
+    fi
+    
     if [[ $? -eq 0 ]]
     then
 		if [[ $_VERBOSE -ge 1 ]]
@@ -304,8 +341,7 @@ ERROR\n" $test_n >> $LOG_FILE"_"$test_n
 				echo $macro
 				cat $ft_printf_diff_file
 			fi
-		elif [[ $_VERBOSE -eq 0 ]]; then
-			printf "\033[32;1m*\033[m"
+		elif [[ $_VERBOSE -eq 0 ]]; then printf "\033[32;1m*\033[m"
 		fi
         echo OK >> $LOG_FILE"_"$test_n
 		OK_NUM=$((OK_NUM + 1))
@@ -317,11 +353,11 @@ ERROR\n" $test_n >> $LOG_FILE"_"$test_n
 			if [[ $_VERBOSE -ge 2 ]]
 			then
     			echo $macro
-        		diff --color=always -u -a --label FT_42 $ft_printf_diff_file\
- --label STDIO $printf_diff_file
+                if [[ -e $ft_printf_diff_file ]] && [[ -e $printf_diff_file ]]; then
+        		    diff --color=always -a -u --label FT_42 $ft_printf_diff_file --label STDIO $printf_diff_file
+                fi
 			fi
-		elif [[ $_VERBOSE -eq 0 ]]; then
-			printf "\033[31;1m!%-6d\033[m" $test_n
+		elif [[ $_VERBOSE -eq 0 ]]; then printf "\033[31;1m!%-6d\033[m" $test_n
 		fi
 		KO_NUM=$((KO_NUM + 1))
 		return 2
@@ -330,9 +366,14 @@ ERROR\n" $test_n >> $LOG_FILE"_"$test_n
 
 function print_summary ()
 {
+    if [[ $test_n -le 0 ]] ; then test_n=$(($test_n + 1))
+    fi
+
 	local percent_ko=$(perl -e 'print '${KO_NUM}'.0 / '${test_n}'.0 * 100' )
+
 	printf "\nSummary : \033[32;1m%6d OK \033[31;1m%6d KO \033[37;1m%6d TOTAL\033[m\t" $OK_NUM $KO_NUM $test_n
-	if [[ ${percent_ko//.*} -lt 40 ]]
+
+	if [[ ${percent_ko//.*} -lt 40 ]] && [[ ${percent_ko//*} != '0' ]]
 	then 
 		printf "\033[32;1mPercent KO : %f%%\033[m\n" $percent_ko
 	else
@@ -343,55 +384,74 @@ function print_summary ()
 function cleanup ()
 {
 	print_summary
-    rm -f $printf_diff_file $printf_exec_file $printf_main_file \
-    $ft_printf_diff_file $ft_printf_exec_file $ft_printf_main_file
+    rm -f $printf_diff_file $printf_exec_file $printf_main_file $ft_printf_diff_file $ft_printf_exec_file $ft_printf_main_file
     exit
 }
 
+function input_files()
+{
+    if [[ -e Makefile ]] || [[ -e makefile ]]; then make; fi
+    FT_PRINTF_HEADER_FILE=$( cd $WD && find ../ -type f -name "ft_printf.h" )
+    if [[ -z $FT_PRINTF_HEADER_FILE ]] ; then echo "ft_printf.h Not Found" ; exit 1
+    fi
+    FT_PRINTF_LIB_FILE=$( find . -name "libftprintf.a" )
+    if [[ -z $FT_PRINTF_LIB_FILE ]] ; then echo "libftprintf.a Not Found" ; exit 1
+    fi
+}
 # ================================== Program ================================= #
 # Arg check
-if [[ $# -ne 1 ]]
-then
-    echo "Usage $0 [Number of tests to run]"
-    exit 1
-fi
+NO_DISPLAY=0
+MAX_TESTS=1
+_VERBOSE=1
+_GLOBAL_MAX_=2
 
-make
+while [[ $1 != "" ]]; do
+    case $1 in
+    -v | --verbose )    shift ; _VERBOSE=$1;;
+    -t | --tests )      shift ; MAX_TESTS=$1;;
+    -c | --complexity ) shift ; _GLOBAL_MAX_=$1;;
+    -d | --no-display ) NO_DISPLAY=1;;
+    *) usage;;
+    esac
+    shift
+done
 
-# Set up cleanup
-trap cleanup SIGINT EXIT SIGSEGV SIGABRT SIGBUS
-# Create Working Directory
-mkdir -p $WD $LOG_DIR
 # Delete old LOGS
 rm -f $LOG_FILE*
+# Create Working Directory
+mkdir -p $WD $LOG_DIR
+# Run make & Check input files
+input_files
 # Write files
 write_main_files
+# Set up cleanup
+trap cleanup SIGINT EXIT SIGSEGV SIGABRT SIGBUS
 # Run the tests
 OK_NUM=0
 KO_NUM=0
 declare -a KO_ARR
-for (( test_n=0;test_n<$1; test_n++ ))
+for (( test_n=0;test_n<$MAX_TESTS; test_n++ ))
 do
     run_test $test_n
 	status=$?
-    if [[ $status -eq 1 ]]
-    then
-        exit 1
-	elif [[ $status -eq 2 ]]
+	if [[ $status -ne 0 ]] 
 	then
 		KO_ARR+=( $LOG_FILE"_"$test_n )
     fi
 done
-print_summary
-rm -f $LOG_DISPLAY_FILE \
-&& echo "cat "${KO_ARR[@]}" | less" > $LOG_DISPLAY_FILE \
-&& chmod 775 $LOG_DISPLAY_FILE
+if [[ KO_NUM -gt 0 ]] && [[ ! $NO_DISPLAY ]]
+then
+    print_summary
+    rm -f $LOG_DISPLAY_FILE \
+    && echo "cat "${KO_ARR[@]}" | less" > $LOG_DISPLAY_FILE \
+    && chmod 775 $LOG_DISPLAY_FILE
 
-while true; do
-    read -p "Do you want to display all the logs of the failed tests in less ?" yn
-    case $yn in
-        [Yy]* ) ./$LOG_DISPLAY_FILE; break;;
-        [Nn]* ) exit;;
-        * ) echo "Please answer yes or no.";;
-    esac
-done
+    while true; do
+        read -p "Do you want to display all the logs of the failed tests in less ?" yn
+        case $yn in
+            [Yy]* ) ./$LOG_DISPLAY_FILE; break;;
+            [Nn]* ) exit;;
+            * ) exit;;
+        esac
+    done
+fi
